+++ etils/epy/adhoc_utils/binary_import.py
@@ -19,8 +19,9 @@ from __future__ import annotations
 from collections.abc import Iterator
 import contextlib
 import functools
+import os
 import sys
-from typing import Any, Optional
+from typing import Optional
 
 from etils.epy import contextlib as epy_contextlib
 from etils.epy import lazy_imports_utils
@@ -65,6 +66,5 @@ def binary_adhoc(
     restrict: None | py_utils.StrOrStrList = None,
     reload: None | py_utils.StrOrStrList = None,
     verbose: bool = False,
-    **kwargs: Any,
 ) -> Iterator[None]:
   yield