+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Changelog follow https://keepachangelog.com/ format.
     *   Add a `edc.ContextStack()` util for a thread-safe / contextvars-aware
         stack.
 *   `epy`:
+    *   `epy.pretty_repr` supports `pydantic` objects.
     *   Support `__setattr__` and `__delattr__` on LazyModules.
     *   Fix `epy.ContextVar` descriptor that were sharing their state.