+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Changelog follow https://keepachangelog.com/ format.
 
 ## [Unreleased]
 
+## [1.13.0] - 2025-07-15
+
 *   `ecolab`:
     *   `adhoc`: Better error message for adhoc invalidate with `epy.reraise`
     *   Fix Digraph `patch_graphviz()`
@@ -510,7 +512,8 @@ Changelog follow https://keepachangelog.com/ format.
 
 <!-- mdlint off(LINK_UNUSED_ID) -->
 
-[Unreleased]: https://github.com/google/etils/compare/v1.12.2...HEAD
+[Unreleased]: https://github.com/google/etils/compare/v1.13.0...HEAD
+[1.12.2]: https://github.com/google/etils/compare/v1.12.2...v1.13.0
 [1.12.2]: https://github.com/google/etils/compare/v1.12.1...v1.12.2
 [1.12.1]: https://github.com/google/etils/compare/v1.12.0...v1.12.1
 [1.12.0]: https://github.com/google/etils/compare/v1.11.0...v1.12.0